Initialization promise on android
On android the initialization code is called in the constructor, which is executed when the JS module is imported.  However, the initialization is asynchronous which means that the JS code may proceed and wind up calling other TTS functions without realizing that the engine is not yet initialized.

This PR moves the initialization to an explicit method which returns a promise which is resolved when initialization is complete.  This way, one can easily chain subsequent calls without any race conditions (and detect when TTS is not available).