dev(DDAPP-9866): do not queue tts when voice selection sample is active
Change isTtsCompleted to isTtsIdle.

Do not start an executer when the request is for a voice selection sample while a voice
selection sample is already active (check for not idle and identical utterance hash) to
avoid additional voice selection samples being played.

Move code from functions only called by audioDone inside audioDone so that the
locking is more visible.

Update gain only when utterances are fully spoken and set min and max values.